updated files to resolve conflicts
