+++ scripts/deep/test/sample_encoded_hotel_data.csv
@@ -1,0 +1,6 @@
+0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31
+-0.16674446,0.1771619,-0.017488815,0.19267826,-0.21439275,0.008447502,0.1482623,-0.24791569,-0.16093448,0.063153364,-0.20009694,-0.119086444,-0.051561084,-0.19371967,-0.20052491,-0.17931525,-0.21760245,0.20078413,0.2113554,-0.30443817,0.08831103,0.25573906,0.0006076073,-0.20432717,-0.096243046,-0.24750361,-0.14462313,-0.22183125,0.15681867,-0.016808808,-0.1712572,0.20596486
+-0.16675071,0.17716138,-0.017491516,0.19270363,-0.21440232,0.008432477,0.14824231,-0.24791418,-0.16092056,0.06316367,-0.20010994,-0.11909133,-0.05158768,-0.19370204,-0.20054565,-0.17930262,-0.21760039,0.20079821,0.21135719,-0.30443424,0.08831142,0.25572172,0.0005839626,-0.20432073,-0.09624641,-0.24751315,-0.14462966,-0.22180943,0.15682364,-0.01679641,-0.17125915,0.20595782
+-0.16678815,0.17719099,-0.017529452,0.19268952,-0.21445578,0.00842463,0.14824446,-0.24787405,-0.16090536,0.063195564,-0.20012687,-0.119070575,-0.051609438,-0.19369748,-0.20048454,-0.17928968,-0.21761017,0.20079198,0.2114035,-0.30441457,0.088288546,0.25573716,0.0005972656,-0.2043307,-0.096262164,-0.24754609,-0.14461148,-0.22176418,0.15682311,-0.01668062,-0.17126745,0.20594616
+-0.16676337,0.17716683,-0.017487004,0.19269265,-0.21440516,0.008433187,0.14826228,-0.24791594,-0.16092135,0.063143164,-0.2001088,-0.11909033,-0.05156208,-0.19370553,-0.20051049,-0.17931248,-0.21761325,0.20078641,0.2113428,-0.30444285,0.08829916,0.25574693,0.00059598155,-0.20432857,-0.096256346,-0.247509,-0.14460768,-0.2218142,0.15682079,-0.01678576,-0.17126152,0.20594807
+-0.16670544,0.17715915,-0.017470865,0.19269192,-0.21437754,0.008440221,0.14825134,-0.24793409,-0.16094619,0.0631086,-0.2001161,-0.11907831,-0.05152729,-0.19373529,-0.20051745,-0.17932236,-0.21760058,0.20079342,0.21132055,-0.30443516,0.088328324,0.25577262,0.0005976214,-0.20433255,-0.096238814,-0.24748026,-0.14464206,-0.22182201,0.1568504,-0.016829312,-0.17123227,0.2059753
